Contract working with SVG images [See description]
But it is 30611 bytes long. And should not be more than 24576 bytes to be deployable on mainnet.
This issue could be resolved by modifying the "MetaCare" text in the svg file, to be as a real text with the Poppins font, instead of (currently) a full image added to the file.